Create a custom_data string with Terraform Ouputs
# Overview
Currently, users have the ability to add `custom_data` to a linux VM in two ways:

1. Passing a local path to a script.
2. Directly inserting code in the `tfvars`

This allows users to manually enter `custom_data` for these machines. But it does __not__ allow users to build a a string for the `custom_data` from the outputs of other Terraform outputs.

# Our Use Case
In our use case, we need to set the following string as the `custom_data`:
```
storage-account=stname, access-key=some-secret, file-share=myfileshare, share-directory=testdirectory
```

# Implementation
The implementation isnt exactly the cleanest. It currently requires a nested `if` statement to mimic an `elif` block. In the future is others want to add other dynamically created string for `custom_data`, they will have to add other another loop.