Fix custom route table on AzureFirewallSubnet
When creating an Azure Firewall, it must be placed on a "special" subnet
named AzureFirewallSubnet. If one wants to add a custom route table to
this subnet, it must contain a "0.0.0.0/0" route to the internet, or
else we get the following error:

"Route Table ktyu-route-test on firewall subnet AzureFirewallSubnet must
have 0.0.0.0/0 route with next hop Internet."

However, even if we have this route in our table, terraform-azurerm-caf
was still throwing the above error as it tries to associate the table
with the subnet before the default route has been added.

This patch fixes this by moving the
azurerm_subnet_route_table_association resource up from the
virtual_network submodule into the root networking.tf module. This way
the dependencies are correctly inferred and Terraform ensures that the
route has been added before trying to create the association.

Fixes https://github.com/aztfmod/terraform-azurerm-caf/issues/868