refactor(storage): replace storage_account_name with storage_share_id
Description:

This commit replaces the use of `storage_account_name` with `storage_share_id` in the `file_share_directory`, `directory`, and `variables` modules. This change is made to improve the referencing mechanism within these modules. The `storage_share_id` provides a more direct and efficient way to reference the Azure Storage Share, eliminating the need for the `storage_account_name`.